# admin-api/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  license:
    name: Apache License
    url: "https://www.apache.org/licenses/LICENSE-2.0"
  title: SaaS Express
servers:
  - url: http://127.0.0.1:8081
security: []
paths:
  /services:
    get:
      summary: Get all services
      operationId: getServices
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: recordsPerPage
          schema:
            type: integer
            default: 25

      responses:
        "200":
          description: A list of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      summary: Create a new service
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{serviceId}:
    get:
      summary: Get a specific service
      operationId: getService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    put:
      summary: Update an existing service
      operationId: updateService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      summary: Delete a service
      operationId: deleteService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Service deleted

  /activity:
    get:
      operationId: "get-activity"
      summary: "Get activity"
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: recordsPerPage
          schema:
            type: integer
            default: 25
      security: []
      responses:
        "200":
          description: activity list response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
        "400":
          description: Invalid response
          content:
            application/json:
              schema:
                type: object

    post:
      summary: "200"
      description: ""
      operationId: "create-activity"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Activity"
          application/yaml:
            schema:
              $ref: "#/components/schemas/Activity"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Activity"
      security: []

      responses:
        "200":
          description: activity response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          description: Invalid response
          content:
            application/json:
              schema:
                type: object
        "422":
          description: Validation exception

  /activity/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: "delete-activity"
      summary: "Delete activity"
      security: []
      responses:
        "200":
          description: activity response
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Activity not found
          content:
            application/json:
              schema:
                type: object

  /tenants:
    get:
      operationId: "get-tenants"
      summary: "Get tenants"
      security: []
      responses:
        "200":
          description: tenant list response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid response
          content:
            application/json:
              schema:
                type: object

    post:
      summary: "200"
      description: ""
      operationId: "create-tenant"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
          application/yaml:
            schema:
              $ref: "#/components/schemas/Tenant"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Tenant"
      security: []

      responses:
        "200":
          description: tenant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid response
          content:
            application/json:
              schema:
                type: object
        "422":
          description: Validation exception

  /tenants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: "delete-tenant"
      summary: "Delete tenant"
      security: []
      responses:
        "200":
          description: tenant response
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                type: object
    get:
      operationId: "get-tenant"
      summary: "Get tenant"
      security: []
      responses:
        "200":
          description: tenant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                type: object
    put:
      summary: "200"
      description: ""
      operationId: "update-tenant"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
          application/yaml:
            schema:
              $ref: "#/components/schemas/Tenant"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Tenant"
      security: []

      responses:
        "200":
          description: tenant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid response
          content:
            application/json:
              schema:
                type: object
        "422":
          description: Validation exception

components:
  schemas:
    Service:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the service
        displayName:
          type: string
        serviceUrl:
          type: string
          description: Upstream endpoint URL for the service
        variants:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/DAGVariant"
        # dag:
        #   type: object
        # visuals:
        #   type: string
    ServiceRequest:
      type: object
      properties:
        displayName:
          type: string
        serviceUrl:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/DAGVariant"
        # dag:
        #   type: object
        # visuals:
        #   type: string
      required:
        - displayName
        - serviceUrl
        # - dag
        # - visuals

    DAGVariant:
      type: object
      properties:
        dag:
          type: object

    Tenant:
      type: object
      properties:
        id:
          type: string
          example: E38A8913ACD5
        displayName:
          type: string
          example: New Tenant
    Activity:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
          example: "{actor} {action} {entity}"
        result:
          type: string
        activityAt:
          type: string
        params:
          type: object
          additionalProperties:
            type: string
          # patternProperties:
          #   "^.*$":
          #     anyOf:
          #       - type: string
          # additionalProperties: false