# admin-api/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// Activity defines model for Activity.
type Activity struct {
	ActivityAt *string            `json:"activityAt,omitempty"`
	Id         *int               `json:"id,omitempty"`
	Message    *string            `json:"message,omitempty"`
	Params     *map[string]string `json:"params,omitempty"`
	Result     *string            `json:"result,omitempty"`
}

// DAGVariant defines model for DAGVariant.
type DAGVariant struct {
	Dag *map[string]interface{} `json:"dag,omitempty"`
}

// Service defines model for Service.
type Service struct {
	DisplayName *string `json:"displayName,omitempty"`

	// Id Unique identifier for the service
	Id *string `json:"id,omitempty"`

	// ServiceUrl Upstream endpoint URL for the service
	ServiceUrl *string                `json:"serviceUrl,omitempty"`
	Variants   *map[string]DAGVariant `json:"variants,omitempty"`
}

// ServiceRequest defines model for ServiceRequest.
type ServiceRequest struct {
	DisplayName string        `json:"displayName"`
	ServiceUrl  string        `json:"serviceUrl"`
	Variants    *[]DAGVariant `json:"variants,omitempty"`
}

// Tenant defines model for Tenant.
type Tenant struct {
	DisplayName *string `json:"displayName,omitempty"`
	Id          *string `json:"id,omitempty"`
}

// GetActivityParams defines parameters for GetActivity.
type GetActivityParams struct {
	Page           *int `form:"page,omitempty" json:"page,omitempty"`
	RecordsPerPage *int `form:"recordsPerPage,omitempty" json:"recordsPerPage,omitempty"`
}

// GetServicesParams defines parameters for GetServices.
type GetServicesParams struct {
	Page           *int `form:"page,omitempty" json:"page,omitempty"`
	RecordsPerPage *int `form:"recordsPerPage,omitempty" json:"recordsPerPage,omitempty"`
}

// CreateActivityJSONRequestBody defines body for CreateActivity for application/json ContentType.
type CreateActivityJSONRequestBody = Activity

// CreateActivityFormdataRequestBody defines body for CreateActivity for application/x-www-form-urlencoded ContentType.
type CreateActivityFormdataRequestBody = Activity

// CreateServiceJSONRequestBody defines body for CreateService for application/json ContentType.
type CreateServiceJSONRequestBody = ServiceRequest

// UpdateServiceJSONRequestBody defines body for UpdateService for application/json ContentType.
type UpdateServiceJSONRequestBody = ServiceRequest

// CreateTenantJSONRequestBody defines body for CreateTenant for application/json ContentType.
type CreateTenantJSONRequestBody = Tenant

// CreateTenantFormdataRequestBody defines body for CreateTenant for application/x-www-form-urlencoded ContentType.
type CreateTenantFormdataRequestBody = Tenant

// UpdateTenantJSONRequestBody defines body for UpdateTenant for application/json ContentType.
type UpdateTenantJSONRequestBody = Tenant

// UpdateTenantFormdataRequestBody defines body for UpdateTenant for application/x-www-form-urlencoded ContentType.
type UpdateTenantFormdataRequestBody = Tenant

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get activity
	// (GET /activity)
	GetActivity(c *gin.Context, params GetActivityParams)
	// 200
	// (POST /activity)
	CreateActivity(c *gin.Context)
	// Delete activity
	// (DELETE /activity/{id})
	DeleteActivity(c *gin.Context, id int)
	// Get all services
	// (GET /services)
	GetServices(c *gin.Context, params GetServicesParams)
	// Create a new service
	// (POST /services)
	CreateService(c *gin.Context)
	// Delete a service
	// (DELETE /services/{serviceId})
	DeleteService(c *gin.Context, serviceId string)
	// Get a specific service
	// (GET /services/{serviceId})
	GetService(c *gin.Context, serviceId string)
	// Update an existing service
	// (PUT /services/{serviceId})
	UpdateService(c *gin.Context, serviceId string)
	// Get tenants
	// (GET /tenants)
	GetTenants(c *gin.Context)
	// 200
	// (POST /tenants)
	CreateTenant(c *gin.Context)
	// Delete tenant
	// (DELETE /tenants/{id})
	DeleteTenant(c *gin.Context, id string)
	// Get tenant
	// (GET /tenants/{id})
	GetTenant(c *gin.Context, id string)
	// 200
	// (PUT /tenants/{id})
	UpdateTenant(c *gin.Context, id string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetActivity operation middleware
func (siw *ServerInterfaceWrapper) GetActivity(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetActivityParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "recordsPerPage" -------------

	err = runtime.BindQueryParameter("form", true, false, "recordsPerPage", c.Request.URL.Query(), &params.RecordsPerPage)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter recordsPerPage: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetActivity(c, params)
}

// CreateActivity operation middleware
func (siw *ServerInterfaceWrapper) CreateActivity(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateActivity(c)
}

// DeleteActivity operation middleware
func (siw *ServerInterfaceWrapper) DeleteActivity(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteActivity(c, id)
}

// GetServices operation middleware
func (siw *ServerInterfaceWrapper) GetServices(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetServicesParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "recordsPerPage" -------------

	err = runtime.BindQueryParameter("form", true, false, "recordsPerPage", c.Request.URL.Query(), &params.RecordsPerPage)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter recordsPerPage: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetServices(c, params)
}

// CreateService operation middleware
func (siw *ServerInterfaceWrapper) CreateService(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateService(c)
}

// DeleteService operation middleware
func (siw *ServerInterfaceWrapper) DeleteService(c *gin.Context) {

	var err error

	// ------------- Path parameter "serviceId" -------------
	var serviceId string

	err = runtime.BindStyledParameterWithOptions("simple", "serviceId", c.Param("serviceId"), &serviceId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter serviceId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteService(c, serviceId)
}

// GetService operation middleware
func (siw *ServerInterfaceWrapper) GetService(c *gin.Context) {

	var err error

	// ------------- Path parameter "serviceId" -------------
	var serviceId string

	err = runtime.BindStyledParameterWithOptions("simple", "serviceId", c.Param("serviceId"), &serviceId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter serviceId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetService(c, serviceId)
}

// UpdateService operation middleware
func (siw *ServerInterfaceWrapper) UpdateService(c *gin.Context) {

	var err error

	// ------------- Path parameter "serviceId" -------------
	var serviceId string

	err = runtime.BindStyledParameterWithOptions("simple", "serviceId", c.Param("serviceId"), &serviceId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter serviceId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateService(c, serviceId)
}

// GetTenants operation middleware
func (siw *ServerInterfaceWrapper) GetTenants(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetTenants(c)
}

// CreateTenant operation middleware
func (siw *ServerInterfaceWrapper) CreateTenant(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateTenant(c)
}

// DeleteTenant operation middleware
func (siw *ServerInterfaceWrapper) DeleteTenant(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteTenant(c, id)
}

// GetTenant operation middleware
func (siw *ServerInterfaceWrapper) GetTenant(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetTenant(c, id)
}

// UpdateTenant operation middleware
func (siw *ServerInterfaceWrapper) UpdateTenant(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateTenant(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/activity", wrapper.GetActivity)
	router.POST(options.BaseURL+"/activity", wrapper.CreateActivity)
	router.DELETE(options.BaseURL+"/activity/:id", wrapper.DeleteActivity)
	router.GET(options.BaseURL+"/services", wrapper.GetServices)
	router.POST(options.BaseURL+"/services", wrapper.CreateService)
	router.DELETE(options.BaseURL+"/services/:serviceId", wrapper.DeleteService)
	router.GET(options.BaseURL+"/services/:serviceId", wrapper.GetService)
	router.PUT(options.BaseURL+"/services/:serviceId", wrapper.UpdateService)
	router.GET(options.BaseURL+"/tenants", wrapper.GetTenants)
	router.POST(options.BaseURL+"/tenants", wrapper.CreateTenant)
	router.DELETE(options.BaseURL+"/tenants/:id", wrapper.DeleteTenant)
	router.GET(options.BaseURL+"/tenants/:id", wrapper.GetTenant)
	router.PUT(options.BaseURL+"/tenants/:id", wrapper.UpdateTenant)
}