# tenant-ui/app/components/graph/types.ts
export interface Service {
  displayName: string;
  variants: Record<string, DAGVariant>;
}

export interface DAGVariant {
  dag: DAG;
}

export interface DAG {
  edges: Edge[];
  nodes: Node[];
  name: string;
  visuals: any;
}

export interface Node {
  id: string;
  x: number;
  y: number;
  label: string;
  action?: string;
}

export interface Edge {
  from: string;
  to: string;
}